[refactor] Sponsor dto id 컬럼 추가 #62
